requirements/base.txt
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow==11.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.8.2  # https://github.com/evansd/whitenoise
redis==5.2.1  # https://github.com/redis/redis-py
hiredis==3.1.0  # https://github.com/redis/hiredis-py
celery==5.4.0  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.7.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==5.0.10  # pyup: < 5.1  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==5.0.0  # https://github.com/jazzband/django-model-utils
django-allauth[mfa]==65.3.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.3  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.16
amqp==5.3.1
anyio==4.4.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
babel==2.16.0
billiard==4.2.1
build==1.2.1
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
cssbeautifier==1.15.1
decorator==5.1.1
distlib==0.3.8
dj-database-url==2.2.0
# django-allauth==0.63.3
django-auto-prefetch==1.4.0
django-ckeditor==6.7.1
django-countries==7.6.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-multiupload==0.6.1
django-phonenumber-field==8.0.0
django-resized==1.0.2
django-stubs-ext==5.1.1
django-taggit==5.0.1
docutils==0.21.2
EditorConfig==0.12.4
executing==2.0.1
Faker==27.0.0
filelock==3.15.4
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httptools>=0.6.1
identify==2.6.0
idna==3.7
imagesize==1.4.1
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.1.0
# mypy==1.10.0
mypy-extensions==1.0.0
nodeenv==1.9.1
nowpayments==1.3.3
packaging==24.1
parso==0.8.4
pathspec==0.12.1
phonenumbers==8.13.48
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
propcache==0.2.0
# psycopg==3.1.19
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycountry==24.6.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pypng==0.20220715.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
regex==2023.12.25
requests==2.32.3
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
# Sphinx==7.3.7
# sphinx-autobuild==2024.4.16
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.38.2
termcolor==2.4.0
text-unidecode==1.3
tqdm==4.66.5
traitlets==5.14.3
twilio==9.3.4
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
# uvicorn==0.30.1
uvicorn-worker==0.2.0
vine==5.1.0
virtualenv==20.26.3
watchdog==4.0.2
wcwidth==0.2.13
websockets==12.0
# Werkzeug==3.0.3
# yarl==1.16.0
